chore(modal): convert tests to RTL
